Stabilise on demand radio tests
