+++ cypress/integration/pages/onDemandRadio/tests.js
@@ -1,3 +1,8 @@
+import {
+  isExpired,
+  dataEndpointOverride,
+} from '../../../support/helpers/onDemandRadioTv';
+
 export default ({ service, pageType }) => {
   describe(`Tests for ${service} ${pageType}`, () => {
     describe('Brand image visible above 400px, not visible below 400px', () => {
@@ -11,6 +16,24 @@ export default ({ service, pageType }) => {
         cy.get('div[class^="ImageContainer"]')
           .find('img')
           .should('not.be.visible');
+      });
+    });
+
+    describe('Audio Player', () => {
+      it('should render an iframe with a valid URL', () => {
+        cy.request(
+          `${Cypress.env('currentPath')}.json${dataEndpointOverride()}`,
+        ).then(({ body: jsonData }) => {
+          const isExpiredEpisode = isExpired(jsonData);
+
+          if (!isExpiredEpisode) {
+            cy.get('iframe').then(iframe => {
+              cy.testResponseCodeAndType(iframe.prop('src'), 200, 'text/html');
+            });
+          } else {
+            cy.log(`Episode is expired: ${Cypress.env('currentPath')}`);
+          }
+        });
       });
     });
   });
